+++ web/app/brep/brep-primitives.js
@@ -2,7 +2,7 @@ import {Point} from './geom/point'
 import {Plane} from './geom/impl/plane'
 import {createPrism, enclose} from './brep-enclose'
 import {AXIS, Matrix3} from '../math/l3space'
-import {Circle} from '../cad/craft/sketch/sketch-model'
+import {Circle} from '../cad/sketch/sketchModel'
 
 export function box(w, h, d, tr) {
   const wh = w * 0.5;
