+++ web/app/cad/craft/brep/cut-extrude.js
@@ -6,8 +6,8 @@ import * as stitching from '../../../brep/stitching'
 import {Loop} from '../../../brep/topo/loop'
 import {incRefCounter} from '../../../brep/topo/topo-object'
 import {Line} from '../../../brep/geom/impl/line'
-import {ReadSketchFromFace} from '../sketch/sketch-reader'
-import {Segment} from '../sketch/sketch-model'
+import {ReadSketchFromFace} from '../../sketch/sketchReader'
+import {Segment} from '../../sketch/sketchModel'
 import {isCurveClass} from '../../cad-utils'
 import {BooleanOperation, combineShells} from './boolean-operation'
